update sdk version to 3.6.1, rename java demo app
