+++ Java/Demo/app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "io.chirp.connectdemoapp"
+        applicationId "io.chirp.sdkdemoapp"
         minSdkVersion 16
         targetSdkVersion 28
         versionCode 1
@@ -18,15 +18,10 @@ android {
     }
 }
 
-
 dependencies {
-    implementation 'io.chirp:chirpsdk:3.6.0-beta'
-
+    implementation 'io.chirp:chirpsdk:3.6.1-SNAPSHOT'
     implementation 'com.android.support:design:28.0.0'
-
     implementation 'com.android.support:appcompat-v7:28.0.0'
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    implementation 'commons-codec:commons-codec:1.11'
 }
