+++ Java/Demo/app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/white"
-    tools:context="io.chirp.connectdemoapp.MainActivity">
+    tools:context="io.chirp.sdkdemoapp.MainActivity">
 
     <RelativeLayout
         android:layout_width="0dp"
