+++ Java/Demo/app/src/main/res/values/colors.xml
