+++ Java/Demo/gradle.properties
