+++ Kotlin/Demo/app/build.gradle
@@ -23,9 +23,8 @@ android {
 }
 
 dependencies {
-    implementation 'io.chirp:chirpsdk:3.6.0-beta'
-    
     implementation 'com.android.support:design:28.0.0'
     implementation 'com.android.support:appcompat-v7:28.0.0'
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'io.chirp:chirpsdk:3.6.1'
 }
